+++ lib/remote_control/data/adapters/android_tv_adapter.dart
@@ -44,6 +44,10 @@ class AndroidTvAdapter implements TvBrandAdapter {
       _transportClient.connect(deviceId: device.id);
 
   @override
+  Future<void> connect({required TvDevice device}) =>
+      _transportClient.connect(deviceId: device.id);
+
+  @override
   Future<void> submitPairingCode({
     required TvDevice device,
     required String pinCode,