+++ lib/remote_control/data/adapters/samsung_adapter.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:one_remote/remote_control/application/transport_log_provider.dart';
 import 'package:one_remote/remote_control/application/transport_log_reader.dart';
 import 'package:one_remote/remote_control/application/tv_brand_adapter.dart';
@@ -51,6 +53,7 @@ class SamsungAdapter implements TvBrandAdapter, TransportLogProvider {
 
   final SamsungTransportClient _transportClient;
   final CommandKeyMap _keyMapper;
+  final Map<String, Future<void>> _connectInFlight = {};
 
   static final _ipv4 = RegExp(r'(\d{1,3}(?:\.\d{1,3}){3})');
 
@@ -76,6 +79,18 @@ class SamsungAdapter implements TvBrandAdapter, TransportLogProvider {
     await _transportClient.requestPairingApproval(
       deviceId: device.id,
       triggerKeyCode: triggerKeyCode,
+    );
+  }
+
+  @override
+  Future<void> connect({required TvDevice device}) {
+    return _connectInFlight.putIfAbsent(
+      device.id,
+      () {
+        final f = _transportClient.connect(deviceId: device.id);
+        unawaited(f.whenComplete(() => _connectInFlight.remove(device.id)));
+        return f;
+      },
     );
   }
 
@@ -129,7 +144,7 @@ class SamsungAdapter implements TvBrandAdapter, TransportLogProvider {
   @override
   Stream<ConnectionState> watchConnectionState(TvDevice device) async* {
     try {
-      await _transportClient.connect(deviceId: device.id);
+      await connect(device: device);
     } catch (_) {
       yield ConnectionState.error;
       return;