+++ test/lib/remote_control/data/multiplexed_tv_connection_state_service_test.dart
@@ -93,6 +93,9 @@ class _RecordingCommandService implements RemoteCommandService {
   Future<void> cancelPairing({required TvDevice device}) async {}
 
   @override
+  Future<void> connect({required TvDevice device}) async {}
+
+  @override
   Future<bool> checkRemoteTextInputReady({required TvDevice device}) async =>
       false;